Create AppConstants
I was debating blocking the actual constants for this, but honestly it's so similar to Flappy Bird that I can't bring myself to care if someone steals the exact jump height of this application. Besides, a lot of the Flappy Bird clones on Google Play have different jump heights and gap constants and they're just setting a level of difficulty or playing to their players' preferences. This page is probably going to change the most out of any one, since I plan on making frequent updates in order to keep my players happy whether I have six or six hundred.